Transaction listed plus bootstrap added

+++ src/main/java/com/example/model/pojo/Wallet.java
@@ -8,12 +8,11 @@ import java.util.*;
 
 public class Wallet implements Comparable< Wallet >, Serializable {
 
-    private int wallettID;
+    private long wallettID;
     private String name;
     private BigDecimal amount;
     private long userId;
     private Set< Category > categories;
-//    private List<Wallet> wallets;
 
 
     public Wallet() {
@@ -27,7 +26,7 @@ public class Wallet implements Comparable< Wallet >, Serializable {
         this.name = name;
     }
 
-    public Wallet(int wallettID, String name, BigDecimal amount, long userId, Set< Category > categories) {
+    public Wallet(long wallettID, String name, BigDecimal amount, long userId, Set< Category > categories) {
         this.wallettID = wallettID;
         this.name = name;
         this.amount = amount;
@@ -39,7 +38,7 @@ public class Wallet implements Comparable< Wallet >, Serializable {
         this.categories = categories;
     }
 
-    public Wallet(int wallettID, String name, BigDecimal amount, int userId) {
+    public Wallet(long wallettID, String name, BigDecimal amount, long userId) {
         this.wallettID = wallettID;
         this.name = name;
         this.amount = amount;
@@ -58,7 +57,6 @@ public class Wallet implements Comparable< Wallet >, Serializable {
     public void setWallettID(int wallettID) {
         this.wallettID = wallettID;
     }
-
 
     public BigDecimal getAmount() {
         return amount;
@@ -82,7 +80,7 @@ public class Wallet implements Comparable< Wallet >, Serializable {
 
     @Override
     public int compareTo(Wallet o) {
-        return this.wallettID - o.wallettID;
+        return (int)(this.wallettID - o.wallettID);
     }
 
 }
